==> 2-process_phospho_data_tyrosene.Rmd <==
---
title: "Ex19 Phospho proteomics"
output: html_document
---

The phospho proteomics pipeline uses the R package [PlexedPiper](https://github.com/vladpetyuk/PlexedPiper). It also requires a connection to the DMS to access data packages.

```{r, echo=F}
library(knitr)
knitr::opts_chunk$set(message=F, warning=F)
t0 <- Sys.time(); print(t0)
```

```{r setup}
library(PlexedPiper)
library(dplyr)
library(PNNL.DMS.utils)
library(Biostrings)
library(tidyverse)


data_package_num <- 4925
data_folder <- "data/phospho_data_tyrosine"
crosstab_prefix <- "ptrc_ex19_tyrosene_4925"
path_to_global_crosstab <- "data/global_data/ptrc_ex19_crosstab_global_gene_original.txt"

dir.create(data_folder)

```



# Study design

Study design information in PlexedPiper is encoded in three tables: fractions, samples, and references. 
These tables can be made using metadata and should be stored on the DMS before processing.


```{r read_study_design}
samples <- read.table("data/samples.txt", sep = "\t", header = T) %>%
  mutate(PlexID = as.character(PlexID))
references <- read.table("data/references.txt",  sep = "\t", header = T) %>%
  mutate(PlexID = as.character(PlexID))

```


# Process MS-GF+ data

MS-GF+ data is processed in several steps. First, read MS-GF+ output from the DMS. (This step can take a while).

```{r read_msgf_data}
msgf_data_path <- file.path(data_folder, "msgfData_original.RData")
if (file.exists(msgf_data_path)) {
  load(msgf_data_path)
} else {
  msnid <- read_msgf_data_from_DMS(data_package_num)
  save(msnid, file=msgf_data_path)
}
show(msnid)

msnid_4924 <- read_msgf_data_from_DMS(4924)
show(msnid_4924)

fractions <- psms(msnid) %>%
  select(Dataset) %>% unique() %>%
  mutate(PlexID = sub("^PTRC_Exp19_Plex_0([12]+)_.*$", "\\1", Dataset))
write.table(fractions, "data/phospho_data/fractions_4925.txt", sep = "\t", quote = F, row.names = F)

```


```{r}
psms_tyrosene <- psms(msnid)
psms_4924 <- psms(msnid_4924)

print("peptides from 4924 shared with 4925")
table(psms_4924$pepSeq %in% psms_tyrosene$pepSeq)

print("Tyrosene in 4924")
table(grepl("Y\\*", psms_4924$peptide))
print("S and T in 4924")
table(grepl("S\\*", psms_4924$peptide) | grepl("T\\*", psms_4924$peptide))

print("Tyrosene in 4925")
table(grepl("Y\\*", psms_tyrosene$peptide))
print("S and T in 4925")
table(grepl("S\\*", psms_tyrosene$peptide) | grepl("T\\*", psms_tyrosene$peptide))

```



## AScore


```{r ascore}
ascore <- get_AScore_results(data_package_num)
msnid <- best_PTM_location_by_ascore(msnid, ascore)
msnid <- apply_filter(msnid, "grepl(\"\\\\*\", peptide)")
show(msnid)

ascore_4924 <- get_AScore_results(4924)
msnid_4924 <- best_PTM_location_by_ascore(msnid_4924, ascore_4924)
msnid_4924 <- apply_filter(msnid_4924, "grepl(\"\\\\*\", peptide)")

psms_tyrosene <- psms(msnid)
psms_4924 <- psms(msnid_4924)

print("peptides from 4924 shared with 4925")
table(psms_4924$pepSeq %in% psms_tyrosene$pepSeq)

print("Tyrosene in 4924")
table(grepl("Y\\*", psms_4924$peptide))
print("S and T in 4924")
table(grepl("S\\*", psms_4924$peptide) | grepl("T\\*", psms_4924$peptide))

print("Tyrosene in 4925")
table(grepl("Y\\*", psms_tyrosene$peptide))
print("S and T in 4925")
table(grepl("S\\*", psms_tyrosene$peptide) | grepl("T\\*", psms_tyrosene$peptide))

```


## Correct peak selection and peptide FDR
We use the target-decoy search strategy method described in [(Elias 2010)](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2922680/). 
Filtering is done first at peptide level, then at protein level, both with max FDR of 1%.


```{r}
## peak selection & Peptide filter
msnid <- correct_peak_selection(msnid)   ## optional for phospho
msnid <- filter_msgf_data_peptide_level(msnid, 0.01)
show(msnid)

```


## Remap accessions

This function remaps UniProt protein accessions to gene symbol.

```{r create_conversion_table}
fst_dms_pth <- path_to_FASTA_used_by_DMS(data_package_num)
fst <- readAAStringSet(fst_dms_pth)

conv <- data.frame(NAME = names(fst)) %>%
  mutate(UNIPROT = str_extract(names(fst), "^(\\S)+"),
         SYMBOL = str_extract(names(fst), "GN=(\\S)+")) %>%
  dplyr::select(-NAME) %>%
  mutate(UNIPROT = sub(".*\\|(.*)\\|.*", "\\1", UNIPROT),
         SYMBOL = sub("GN=", "", SYMBOL))

msnid <- remap_accessions_uniprot_to_gene(msnid,
                                          conversion_table = conv)

show(msnid)
table(is.na(msnid$accession))

psms(msnid) %>% distinct(Protein, accession) %>%
  group_by(is.na(accession)) %>%
  summarize(n=n()) %>% print()

msnid <- apply_filter(msnid, "!is.na(accession)")
show(msnid)
```


## Parsimonious inference

To reduce number of protein identifications, we use a parsimonious inference algorithm described in 
[(Zhang et al. 2007)](https://www.ncbi.nlm.nih.gov/pmc/articles/PMC2810678/).


```{r parsimonious_inference}
global_crosstab <- read.table(path_to_global_crosstab, sep = "\t")
msnid <- infer_parsimonious_accessions(msnid, unique_only=FALSE,
                                       prior = rownames(global_crosstab))
show(msnid)

```


## Remove decoys


```{r apply_decoy_filter}
msnid <- apply_filter(msnid, "!isDecoy")
show(msnid)

```


## Map mod sites


```{r map_mod_sites}
path_to_FASTA <- path_to_FASTA_used_by_DMS(data_package_num)
path_to_FASTA <- gsub("\\\\","/",path_to_FASTA)
path_to_FASTA_gene <- remap_accessions_uniprot_to_gene_fasta(path_to_FASTA)

library(Biostrings)
fst <- readAAStringSet(path_to_FASTA_gene)
msnid <- map_mod_sites(msnid, fst, 
                       accession_col = "accession", 
                       peptide_mod_col = "peptide", 
                       mod_char = "*",
                       site_delimiter = "lower")
# msnid <- apply_filter(msnid, "!is.na(PepLocFirst)")
head(psms(msnid))

saveRDS(msnid, file=file.path(data_folder, "msgfData_filtered.RDS"))

```


# Process MASIC data

Output from the MASIC software is read from DMS, then filtered by inteference score.


```{r read_masic}
masic_data_path <- file.path(data_folder, "masicData_original.RData")

if (file.exists(masic_data_path)) {
  load(masic_data_path)
} else {
  masic_data <- read_masic_data_from_DMS(data_package_num,
                                       interference_score = TRUE)
  save(masic_data, file=masic_data_path)
}

nrow(masic_data)

```


```{r filter_masic}
masic_data <- filter_masic_data(masic_data,
                                interference_score_threshold = 0.5,
                                s2n_threshold = 0)
saveRDS(masic_data, file=file.path(data_folder, "masicData_filtered.RDS"))

nrow(masic_data)
```


# Create crosstab

The quantitative crosstab combines MS/MS identifications with reporter ion intensities. 
Abundances are taken relative to the reference channel and then log-transformed.


```{r create_phospho_site_crosstab}
aggregation_level <- c("SiteID")
quant_cross_tab <- create_crosstab(msnid, 
                                   masic_data, 
                                   aggregation_level, 
                                   fractions, samples, references)
dim(quant_cross_tab)

write.table(quant_cross_tab, file=file.path(data_folder, paste(crosstab_prefix, "crosstab_tyrosine_4925_phospho_SiteID_original.txt", sep="_")),
            quote=F, sep="\t")
```


```{r}
## Original global crosstab, using to get sample coefficients for phospho.
global_crosstab <- read.table(path_to_global_crosstab, sep = "\t", header = T)
colnames(global_crosstab) <- sub("\\.", "/", colnames(global_crosstab))
colnames(global_crosstab) <- gsub("\\.", " ", colnames(global_crosstab))
global_crosstab <- global_crosstab[, colnames(quant_cross_tab)]

table(colnames(global_crosstab) == colnames(quant_cross_tab))
global_sample_coefs <- apply(global_crosstab, 2, mean, na.rm = T)
quant_cross_tab <- sweep(quant_cross_tab, 2, global_sample_coefs, FUN = '-')  ##USING GLOBAL COEFFCIENTS
quant_cross_tab <- sweep(quant_cross_tab, 1, apply(quant_cross_tab, 1, mean, na.rm = T), FUN = '-')

write.table(quant_cross_tab, file=file.path(data_folder, paste(crosstab_prefix, "crosstab_tyrisine_4925_phospho_SiteID_median_centered.txt", sep="_")),
            quote=F, sep="\t")
```



```{r}


```
